feat: move LogService to the other folder
